Add Redux Toolkit and OMDB API service integration
In this commit, we integrated Redux toolkit for state management and designed services that interface with the OMDB API for serial and movie data fetching. We created 'omdbApi' and 'posterApi' services that provide query endpoints to access movie data based on the title and IMDB ID. These services are utilized in the new 'Movie' component to query movie data.

Also updated the store, adding these API slices to the global reducer and the necessary middleware to our application's redux store. A StoreProvider component was also created to wrap our application with the Redux Provider.

Add .env.local to .gitignore to prevent environmental variables from being committed to the repository.

Additional minor changes include updates to the 'page.tsx' and 'layout.tsx' files to incorporate our new Movie and StoreProvider components. Lastly, unused styles in the 'globals.css' have been cleared out.