Remove posterApi.ts and optimize redux slices and stores
Deleted posterApi.ts as it was essentially a duplicate with omdbApi.ts. This cleans up code redundancy in the project structure and optimizes state management in redux. Updates were made to store.ts to reflect this deletion.

Additional new files were created and existing ones updated to sophisticate the handling and display of API requested movie data. Moreover, integration of use-debounce in package.json is to facilitate efficient searching by limiting rate of API calls.

Furthermore, implemented a NotFound page to handle cases when the search yields no results, improving user experience and error handling. Also, the Movie component was refactored to handle different states of the data - loading, error or success.

Added new dependencies such as use-debounce in package.json and updated pnpm-lock.yaml file accordingly. This is to make the application more performant and provide a smoother user experience.

Updated README.md to include additional resources and new functionalities such as use-debounce, zod, and clx for ease of accessibility for new contributors.