Add Vercel Analytics package to project dependencies
Added "@vercel/analytics" package version 1.1.1 to the project dependencies listed in the package.json and pnpm-lock.yaml files. This was done to enable the monitoring and collection of application usage data, which helps in improving the application based on user interaction patterns. This additional analytics package will aid in identifying potential areas of improvement in our application.